https://github.com/pendulum-chain/spacewalk/pull/365#discussion_r1259884946,
