+++ clients/wallet/Cargo.toml
@@ -11,6 +11,7 @@ testing-utils = []
 [dependencies]
 async-trait = "0.1.40"
 futures = "0.3.5"
+hex = "0.4.3"
 parity-scale-codec = "3.0.0"
 rand = "0.8.5"
 reqwest = { version = "0.11", features = ["json"] }
@@ -30,3 +31,4 @@ primitives = { package = "spacewalk-primitives", path = "../../primitives"}
 [dev-dependencies]
 mockall = "0.8.1"
 serial_test = "0.9.0"
+
