+++ clients/wallet/src/horizon/mod.rs
@@ -1,0 +1,13 @@
+pub mod responses;
+mod serde;
+mod traits;
+
+mod horizon;
+#[cfg(test)]
+mod tests;
+
+pub use horizon::*;
+pub use traits::HorizonClient;
+
+// todo: change to Slot
+pub type Ledger = u32;
