+++ primitives/Cargo.toml
@@ -10,6 +10,7 @@ codec = { package = "parity-scale-codec", version = "3.0.0", default-features = 
 scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
 serde = { version = "1.0.130", default-features = false, optional = true }
 base58 = { version = "0.2.0" }
+hex = { version = "0.4.3", optional = true }
 
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.40", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.40", default-features = false }
@@ -29,4 +30,5 @@ std = [
     "sp-runtime/std",
     "frame-support/std",
     "substrate-stellar-sdk/std",
+    "hex"
 ]
