+++ primitives/src/lib.rs
@@ -1,6 +1,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![allow(non_upper_case_globals)]
 
+extern crate core;
+
 use base58::ToBase58;
 use codec::{Decode, Encode, MaxEncodedLen};
 use frame_support::error::LookupError;
@@ -30,7 +32,8 @@ use stellar::{
 pub use substrate_stellar_sdk as stellar;
 use substrate_stellar_sdk::{
 	types::{OperationBody, SequenceNumber},
-	ClaimPredicate, Claimant, Memo, MuxedAccount, Operation, TransactionEnvelope,
+	ClaimPredicate, ClaimableBalanceId, Claimant, Memo, MuxedAccount, Operation, StellarSdkError,
+	TransactionEnvelope, XdrCodec,
 };
 
 #[cfg(test)]
@@ -107,11 +110,69 @@ impl<AccountId, CurrencyId: Copy> VaultId<AccountId, CurrencyId> {
 pub type StellarPublicKeyRaw = [u8; 32];
 
 #[cfg(feature = "std")]
-pub fn convert_stellar_public_key_to_encoded(
-	public_key: &PublicKey,
-) -> Result<String, std::str::Utf8Error> {
-	let encoded = public_key.to_encoding();
-	std::str::from_utf8(&encoded).map(|x| x.to_string())
+pub trait StellarTypeToString<T, E: From<std::str::Utf8Error>> {
+	fn as_encoded_string(&self) -> Result<String, E>;
+
+	fn as_a(&self) -> Result<T, StellarSdkError>;
+}
+
+#[cfg(feature = "std")]
+impl<E: From<std::str::Utf8Error>> StellarTypeToString<Self, E> for PublicKey {
+	fn as_encoded_string(&self) -> Result<String, E> {
+		let x = self.to_encoding();
+		let str = std::str::from_utf8(&x).map_err(E::from)?;
+		Ok(str.to_string())
+	}
+
+	fn as_a(&self) -> Result<PublicKey, StellarSdkError> {
+		Ok(self.clone())
+	}
+}
+
+#[cfg(feature = "std")]
+impl<E: From<std::str::Utf8Error>> StellarTypeToString<PublicKey, E> for &str {
+	fn as_encoded_string(&self) -> Result<String, E> {
+		Ok(self.to_string())
+	}
+
+	fn as_a(&self) -> Result<PublicKey, StellarSdkError> {
+		PublicKey::from_encoding(self)
+	}
+}
+
+#[cfg(feature = "std")]
+impl<E: From<std::str::Utf8Error>> StellarTypeToString<PublicKey, E> for Vec<u8> {
+	fn as_encoded_string(&self) -> Result<String, E> {
+		let str = std::str::from_utf8(self).map_err(E::from)?;
+		Ok(str.to_string())
+	}
+
+	fn as_a(&self) -> Result<PublicKey, StellarSdkError> {
+		PublicKey::from_encoding(self)
+	}
+}
+
+#[cfg(feature = "std")]
+impl<E: From<std::str::Utf8Error>> StellarTypeToString<Self, E> for ClaimableBalanceId {
+	fn as_encoded_string(&self) -> Result<String, E> {
+		let xdr = self.to_xdr();
+		Ok(hex::encode(xdr))
+	}
+
+	fn as_a(&self) -> Result<Self, StellarSdkError> {
+		Ok(self.clone())
+	}
+}
+
+#[cfg(feature = "std")]
+impl<E: From<std::str::Utf8Error>> StellarTypeToString<ClaimableBalanceId, E> for &str {
+	fn as_encoded_string(&self) -> Result<String, E> {
+		Ok(self.to_string())
+	}
+
+	fn as_a(&self) -> Result<ClaimableBalanceId, StellarSdkError> {
+		todo!()
+	}
 }
 
 #[cfg(feature = "std")]
